opt: add rule to push groupby into join as limit
This commit adds a new normalization rule, `PushGroupByIntoConstJoin`,
which replaces a GroupBy on a Join operator with a Limit on the right
input of the join. This is possible when the grouping columns form a
key over the left input of the join and all aggregations are ConstAgg-like.
See the rule comment for details.

Epic: CRDB-35243

Release note: None