opt: make WITH binding copying explicit in CopyAndReplace
This commit adds a `addWithBindings` argument to `CopyAndReplace` which
specifies whether WITH bindings from the source memo should be copied to
the destination memo (without applying the replace function). This is
useful for subqueries within routines and apply-joins, which require
copying a sub-expression embedded in the main query that might reference
a WITH binding elsewhere in the main query. Failing to copy the bindings
can result in a "no binding for WithID 1" panic.

There are two cases where callers do not need to set `addWithBindings`:
1. The entire root expression is being copied. In this case, all WITH
   expressions are being copied anyway, along with their bindings.
2. The sub-expression is known not to include any references to "outer"
   WITH bindings.

This change simplifies usage of `CopyAndReplace` for sub-expressions,
and also makes it less error-prone.

Epic: None

Release note: None