kv: add support for best-effort per-scan request limit
This commit adds support for limiting the number of KVs returned for
individual `Scan` and `ReverseScan` requests in a batch. This is a
best effort limit to simplify splitting scans that span multiple
ranges, as well as simplify handling for SQL rows that map to multiple
KVs. Users of the limit should be prepared to handle extra KVs.
Following commits will use this limit to optimize lookup joins which
only need a limited number of rows from each lookup.

NOTE: unlike the batch-level key limit, the per-scan key limit does
not include resume spans in the results. This significantly simplifies
handling of requests that span multiple ranges, and is sufficient for
most use cases.

Informs #62471

Release note: None